Address assertions not working properly in while loops.
For some reason, assertions do not fail properly if they are in a
while loop with shunit2.  Because of this, I've done something a
bit absurd.  We're reading a text file with a while and curling
based on lines, but then doint all our assertions in a for loop.
See previous commit for example where this is an issue.